fastapi_recom/requirements.txt
fastapi[all]===0.111.0
fastapi-cache2==0.2.1
redis==5.2.1
alembic==1.7.7
alembic-autogenerate-enums==0.1.2
asyncpg==0.28.0
black==22.3.0
cachetools==5.3.2
python-dotenv==1.0.1
python-json-logger==2.0.4
SQLAlchemy==2.0.36
faststream[rabbit]==0.5.12
apscheduler==3.11.0
tenacity
uvicorn==0.23.1
slowapi==0.1.8
aio-pika==9.1.5
asyncpg==0.28.0
greenlet>=2.0
pydantic==2.4.1
pydantic-core==2.10.1
pydantic-settings==2.0.3
numpy==1.26.0
scikit-learn==1.7.0
